Add new-style video upload
The previous upload methods required to read the entire video file in
memory during upload; also, it required periodical checks on the file
upload status. Here, we choose to consider that file storage will be
local; thus, file transfer is synchronous. It greatly simplifies things.
In particular, many methods, tests and tasks need to be removed.

For now, we remain compatible with the previous upload method, but this
is going to change.