Need to add a migration because of "Bokmal" typo
Django fixed a typo in a Norwegian dialect: "Bokmal" -> "Bokmål".

Bokmål (literally "book tongue") is an official written standard for the
Norwegian language, alongside Nynorsk. Bokmål is the preferred written
standard of Norwegian for 85% to 90% of the population in Norway, and is
most used by people who speak Standard Østnorsk.
(source: Wikipedia)