Force fake-factory version
In `factory-boy==2.7.0`, there is this requirement `fake-factory>=0.5.0`

But the last version of `fake-factory`, which is `9999.9.9`, is just
telling that this lib does not exist anymore, and it is this version
that is installed on a fresh new venv.

So forcing the version of `fake-factory` to be the last know
working/existing version is the minimal update to make it work.